A major remodelling of the classes functions
