+++ app/room.py
@@ -1,55 +1,22 @@
-import abc
-from amity import Amity
+class Room(object):
 
-class Room(Amity):
-    full_room  = {} # name of the room as the key and names of the occupants as the kvalues
-    not_full_room = {}
     '''
     Models all the information of the rooms that the office and Living space will
     inherit from
     '''
-    def __init__(self, room_name, room_type):
+    def __init__(self, room_name, room_type, max_occupants=0):
         self.room_name = room_name
         self.room_type = room_type
+        self.max_occupants = max_occupants
 
-    @abc.abstractmethod
-    def set_room_max_occupants(self):
-        pass
-
-    def create_room(self):
-        '''
-        Creates a room in Amity
-        '''
-        pass
-
-    def validate_room_type(self):
-        '''
-        Validates that the room is either #Office or #LivingSpace
-        '''
-        pass
-
-    def validate_wing(self, wing):
-        '''
-        if it is a living space, then validates that it is either
-        male_wing or female_wing
-        '''
-        self.wing = wing
-        pass
 
 class Office(Room):
 
-    def room_max_occupants(self):
-        '''
-        set maximum occupants of one office to 6 people
-        '''
-        self.max_occupants = 6
-        return self.max_occupants
+    def __init__(self, room_name):
+        super(Office, self).__init__(room_name, room_type='office') #max_occupants = 6)
+
 
 class LivingSpace(Room):
 
-    def room_max_occupants(self):
-        '''
-        set maximum occupants of one living space to 4 people
-        '''
-        self.max_occupants = 4
-        return self.max_occupants
+    def __init__(self, room_name):
+        super(LivingSpace, self).__init__(room_name, room_type='livingspace') #max_occupants=4)
